release-package/pyproject.toml
[build-system]
requires = ["setuptools>=61.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sphinx-translation-button"
version = "1.1.0"
description = "Manage translations of a book using a launchbuttons"
readme = "README.md"
dependencies = ["sphinx", "PyYAML"]
requires-python = ">=3.9"
